requirements.txt
pytest ~= 7.3.1
selenium ~= 4.9.0
selene ~= 2.0.0rc2
request ~= 2.28.2
xlrd
openpyxl
pypdf
